+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-12-23
+### Added
+- Implemented Win API for event subscription implementations
+- historical flag to `listen_events`
+
+### Changed
+- updated `listen_events` from `0.1.0` to `0.2.0`
+
 ## [0.2.0] - 2019-12-22
 ### Added
 - Enhanced XML to JSON conversion